Package found for Grid Formation
Masonry seems to work pretty well to get the grid assymetrically setup how we wanted, however it does have some bugs that I am having trouble resolving, and may be doing too much of the work. I'm going to keep it on in the master branch and make an alternate branch for proceding without it.